requirements.txt
streamlit_option_menu==0.3.2
streamlit==1.22.0
pandas==2.0.2
mysql-connector-python==8.0.33pi
